# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DomHunter API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from "./domainStatus";
import type { Metrics } from "./metrics";

export interface Domain {
  id: string;
  name: string;
  sld: string;
  tld: string;
  status: DomainStatus;
  source?: string | null;
  auctionUrl?: string | null;
  auctionEndAt?: string | null;
  currentBid?: number | null;
  bidCount?: number | null;
  createdAt: string;
  updatedAt: string;
  metrics?: Metrics | null;
}